FTP Normed

+++ ftp/libft/ft_printf/ft_itoa_base.c
@@ -3,22 +3,22 @@
 /*                                                        :::      ::::::::   */
 /*   ft_itoa_base.c                                     :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: gbourgeo <marvin@42.fr>                    +#+  +:+       +#+        */
+/*   By: gbourgeo <gbourgeo@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2017/08/15 22:06:38 by gbourgeo          #+#    #+#             */
-/*   Updated: 2017/08/15 22:36:31 by gbourgeo         ###   ########.fr       */
+/*   Updated: 2020/02/20 21:35:29 by gbourgeo         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "ft_printf.h"
 #include "libft.h"
 
-char			*ft_itoa_base(ULL nb, ULL base)
+char			*ft_itoa_base(unsigned long long nb, unsigned long long base)
 {
-	char		*b;
-	ULL			size;
-	ULL			i;
-	char		*ret;
+	char				*b;
+	unsigned long long	size;
+	unsigned long long	i;
+	char				*ret;
 
 	b = "0123456789abcdef";
 	size = (nb == 0) ? 1 : 0;
